Separated Config and Datastore
