+++ appmock/src/main/java/land/fx/app/MainActivity.kt
@@ -2,7 +2,7 @@ package land.fx.app
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import land.fx.wnfslib.Bridge.initRustLogger
+import land.fx.wnfslib.Fs.initRustLogger
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
