+++ lib/src/main/java/land/fx/wnfslib/Datastore.java
@@ -1,0 +1,7 @@
+package land.fx.wnfslib;
+
+public interface Datastore {
+    byte[] put(byte[] data, long codec);
+
+    byte[] get(byte[] cid);
+}
